v1.2 - fixed some major issues related to sensor get stuck and sensitivity

+++ livolo1.X/capsensor.h
@@ -11,7 +11,7 @@ extern "C" {
 uint16_t cap_raw[2];
 uint16_t cap_rolling_avg[2];
 uint16_t cap_frozen_avg[2];
-uint8_t cap_cycles[2];
+uint16_t cap_cycles[2];
 
 
 /*
